refactor: queryfn 공용 폴더에 정리
