+++ src/pages/brand/magazine-list.tsx
@@ -1,7 +1,7 @@
 import {useQuery} from '@tanstack/react-query'
 import BrandListContainer from './brand-list-container'
-import {fetchBrandMagazines} from './queryfn'
 import Magazine from './magazine'
+import { fetchBrandMagazines } from 'src/store/server/brand/queries'
 
 interface proptype {
   brandId: string
