+++ src/pages/my-page/feed-section/feed-section.tsx
@@ -1,10 +1,10 @@
 import {Stack} from '@mui/material'
 import FeedChip from '../base/feed-chip'
-import {fetchReviews} from './queryfn'
 import {useQuery} from '@tanstack/react-query'
 import {useState} from 'react'
 import {ShowBenchmarkType} from './type'
 import BranchoutReviews from './branch-out-reviews'
+import { fetchReviewPage } from 'src/store/server/reviews/queries'
 
 const FeedSection = () => {
   const [page, setPage] = useState<number>(1)
@@ -16,7 +16,7 @@ const FeedSection = () => {
   }
 
   const {data: reviews} = useQuery(['reviews', {page: page}], () =>
-    fetchReviews(page),
+    fetchReviewPage('',page,1000)
   )
 
   return (
